requirements.txt
chromedriver-binary==89.0.4389.23.0
Pillow==8.1.0
selenium==3.141.0
urllib3==1.26.3
